MGS: Implement "get ignition state of all SPs" endpoint (#728)
* Custom Ser/De impls for `SerialConsole` packet to allow variable lengths

* Implement "get ignition state of all SPs" endpoint

* minor cleanup/tweaks

* typo fix


* typo fixes

* Expand on comments regarding packet max sizes

+++ gateway/src/sp_comms/serial_console_history.rs
@@ -157,7 +157,7 @@ enum Slot {
     // an in-order serial console packet
     Valid {
         offset: u64,
-        len: u8,
+        len: u16,
         data: [u8; SerialConsole::MAX_DATA_PER_PACKET],
     },
 }
